fs2lib rename to interopfs2
