+++ modules/fs2/src/smithy4s/interopfs2/package.scala
@@ -2,11 +2,11 @@ package smithy4s
 
 import _root_.fs2.Stream
 
-package object fs2lib {
+package object interopfs2 {
 
-  type XmlByteStreamEncoder[F[_], A] =
+  type ByteStreamEncoder[F[_], A] =
     smithy4s.codecs.Writer[Any, Stream[F, Byte], A]
-  type XmlByteStreamDecoder[F[_], A] =
+  type ByteStreamDecoder[F[_], A] =
     smithy4s.codecs.Reader[F, Stream[F, Byte], A]
 
 }
